requirements.txt
anyio==3.6.2
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
fastapi==0.95.0
fonttools==4.39.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
pycodestyle==2.10.0
pydantic==1.10.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==20.2
pytz==2023.3
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
tomli==2.0.1
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.21.1
zipp==3.15.0
